Added a trigger for press event. (tap and hold for 750ms).
Twisted swipe logic to closer to native. (smaller travel to trigger it)

Fixed (put back) negative margin on <a> to make an entire <li> clickable. (it left a border of 10px before the fix.)

Removed extra left border in splitscreen.

Added a 1px ridge top-border for toolbar and navibar (closer to native look).